mamaput_api/requirements.txt
aniso8601==9.0.1
black==24.8.0
blinker==1.8.2
cachelib==0.13.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
coverage==7.6.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-HTTPAuth==4.8.0
Flask-JWT-Extended==4.6.0
Flask-Mail==0.10.0
Flask-OAuthlib==0.9.6
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
google-api-core==2.22.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
greenlet==3.1.1
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
marshmallow==3.22.0
mypy-extensions==1.0.0
oauthlib==2.1.0
packaging==24.1
pathspec==0.12.1
paystackapi==2.1.3
platformdirs==4.3.2
pluggy==1.5.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyJWT==2.9.0
pyparsing==3.2.0
pytest==8.3.2
pytest-cov==4.1.0
pytest-sugar==1.0.0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
requests-oauthlib==1.1.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.36
termcolor==2.4.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
Werkzeug==3.0.6
